Move css style related to scwidgets.css_style level
The css style related functions are not essential for creating a
exercise notebook so we move it one level higher. For that we rename the
`_css_style.py` to `css_style.py`.
